Added authorization to destroy method

+++ app/Http/Controllers/ServerController.php
@@ -129,6 +129,8 @@ class ServerController extends Controller
      */
     public function destroy(Server $server)
     {
+      $this->authorize('delete', $server);
+
       $server->delete();
 
       return redirect('servers');
